We completed the user profile DTO and we add a new logic for the loader of likes

+++ src/main/java/com/ecolink/spring/dto/ClientProfileDTO.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 public class ClientProfileDTO {
 
     private String name;
-    private Integer level;
+    private Long level;
     private List<ClientMissionDTO> completedMissions;
     private List<PostRelevantDTO> listLikePost;
